Setup shopify as a library
Cabal shopify setup as library
Added Network.API prefix to module names
Moved src files to src/Network/API
