+++ shopify.cabal
@@ -1,97 +1,58 @@
--- Initial InTheKnow.cabal generated by cabal init.  For further
--- documentation, see http://haskell.org/cabal/users-guide/
-
--- The name of the package.
 name:                shopify
-
--- The package version.  See the Haskell package versioning policy (PVP)
--- for standards guiding when and how versions should be incremented.
--- http://www.haskell.org/haskellwiki/Package_versioning_policy
--- PVP summary:      +-+------- breaking API changes --                   | | +----- non-breaking API additions
---                   | | | +--- code changes with no API change
 version:             0.0.0.1
-
--- A short (one-line) description of the package.
--- synopsis:
-
--- A longer description of the package.
--- description:
-
--- URL for the project homepage or repository.
 homepage:
-
--- The license under which the package is released.
 license:             AllRightsReserved
-
--- The file containing the license text.
 license-file:        LICENSE
-
--- The package author(s).
 author:              Aaron Levin
-
--- An email address to which users can send suggestions, bug reports, and
--- patches.
 maintainer:          aaron@weirdcanada.com
-
--- A copyright notice.
--- copyright:
-
 category:            Web
-
 build-type:          Simple
-
--- Extra files to be distributed with the package, such as examples or a
--- README.
--- extra-source-files:
-
--- Constraint on the version of Cabal needed to build this package.
 cabal-version:       >=1.10
 
 
-executable shopify
-  -- .hs or .lhs file containing the Main module.
-  main-is: Api.hs
+library
+  exposed-modules:
+        Network.API.Shopify
+        Network.API.Shopify.Types
+        Network.API.Shopify.Client
+    
+  build-depends: 
+        base >=4.6 && <4.8
+      , aeson
+      , bytestring
+      , conduit
+      , comonad
+      , containers
+      , free
+      , http-conduit
+      , mtl
+      , resourcet
+      , scientific
+      , text
+      , time
+      , unordered-containers
 
-  -- Modules included in this executable, other than Main.
-  other-modules: Shopify, Shopify.Types
-
-  -- LANGUAGE extensions used by modules in this package.
-  other-extensions:    OverloadedStrings
-                     , GADTs
-
-  -- Other library packages from which modules are imported.
-  build-depends: base >=4.6 && <4.8
-               , aeson
-               , bytestring
-               , conduit
-               , comonad
-               , containers
-               , free
-               , http-conduit
-               , mtl
-               , resourcet
-               , scientific
-               , text
-               , time
-               , unordered-containers
-
-  -- Directories containing source files.
-  hs-source-dirs: src
-
-  -- Base language which the package is written in.
-  default-language:    Haskell2010
+  hs-source-dirs: 
+        src
+  default-language:    
+        Haskell2010
 
 test-suite shopify-tests
-  hs-source-dirs: test
-  main-is: Main.hs
-  Type: exitcode-stdio-1.0
-  default-language: Haskell2010
-  build-depends: base
-               , shopify -any
-               , QuickCheck
-               , HUnit
-               , test-framework
-               , test-framework-hunit
-               , test-framework-quickcheck2
-               , raw-strings-qq
+  hs-source-dirs: 
+        test
+  main-is: 
+        Main.hs
+  Type: 
+        exitcode-stdio-1.0
+  default-language: 
+        Haskell2010
+  build-depends: 
+        base
+      , shopify -any
+      , QuickCheck
+      , HUnit
+      , test-framework
+      , test-framework-hunit
+      , test-framework-quickcheck2
+      , raw-strings-qq
