+++ src/Network/API/Shopify.hs
@@ -1,0 +1,5 @@
+module Network.API.Shopify (
+  module Network.API.Shopify.Types
+  ) where
+
+import Network.API.Shopify.Types
