+++ src/Network/API/Shopify/Types.hs
@@ -1,0 +1,17 @@
+module Network.API.Shopify.Types ( 
+    Image(..)
+  , InventoryPolicy(..)
+  , Metafield(..)
+  , MetafieldType(..)
+  , Option(..)
+  , Product(..)
+  , Variant(..)
+  ) where
+
+import Network.API.Shopify.Types.Image
+import Network.API.Shopify.Types.InventoryPolicy
+import Network.API.Shopify.Types.Metafield
+import Network.API.Shopify.Types.MetafieldType
+import Network.API.Shopify.Types.Option
+import Network.API.Shopify.Types.Product
+import Network.API.Shopify.Types.Variant
