--- src/Shopify.hs
@@ -1,5 +1,0 @@
-module Shopify (
-  module Shopify.Types
-  ) where
-
-import Shopify.Types
