--- src/Shopify/Types.hs
@@ -1,17 +1,0 @@
-module Shopify.Types ( 
-    Image(..)
-  , InventoryPolicy(..)
-  , Metafield(..)
-  , MetafieldType(..)
-  , Option(..)
-  , Product(..)
-  , Variant(..)
-  ) where
-
-import Shopify.Types.Image
-import Shopify.Types.InventoryPolicy
-import Shopify.Types.Metafield
-import Shopify.Types.MetafieldType
-import Shopify.Types.Option
-import Shopify.Types.Product
-import Shopify.Types.Variant
